requirements.txt
# FastAPI and WebSocket dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# HTTP client and async file operations
aiohttp==3.9.1
aiofiles==23.2.1

# Data validation and models
pydantic>=2.8.0

# Environment variables
python-dotenv==1.0.0

# Audio processing (if needed for future enhancements)
# python-multipart==0.0.6  # For file uploads if needed

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1